// frontend/src/app/api-client/model/accessibilityIssueResponse.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AccessibilityIssueResponse { 
    id?: number;
    scanId?: number;
    ruleCode?: string;
    message?: string;
    severity?: string;
    wcagCriterion?: string;
    pageUrl?: string;
    htmlSnippet?: string;
    selector?: string;
    recommendation?: string;
    helpUrl?: string;
}