// frontend/src/app/api-client/model/scanResponse.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ScanResponse { 
    id?: number;
    projectId?: number;
    status?: string;
    startedAt?: string;
    finishedAt?: string;
}